removed dtype=float32 argument

+++ utils/geometry.py
@@ -110,5 +110,5 @@ def get_euler_matrix(dihedral, twist, sweep):
         [c2*c3, -s2, c2*s3],
         [s1*s3+c1*c3*s2, c1*c2, c1*s2*s3-c3*s1],
         [c3*s1*s2-c1*s3, c2*s1, c1*c3+s1*s2*s3]
-        ], dtype='float32')
+        ])
     return euler_matrix
